Add unit tests for StringUtils; fix hex bug

+++ src/cpp/DisplayProberLib.cpp
@@ -202,7 +202,7 @@ std::string BuildPrimaryPortKey(const gdi::GdiDisplayConfig& config) {
     return {};
   }
 
-  auto tp_id = "0x" + IntsToHex(static_cast<uint32_t>(config.target_path_id));
+  auto tp_id = "0x" + IntsToHex(config.target_path_id);
 
   return std::format("acd_ppk:gpu_id={};tp_id={}", gpu_identity, tp_id);
 }
